Move all validators to separate file
