+++ backend/routes/api/checkIns.js
@@ -1,12 +1,10 @@
-const express = require('express');
+const router = require('express').Router();
 const asyncHandler = require('express-async-handler');
 const { check } = require('express-validator');
 
 const { restoreUser } = require('../../utils/auth');
 const { handleValidationErrors } = require('../../utils/validation');
 const { Venue, CheckIn } = require('../../db/models');
-
-const router = express.Router();
 
 router.get('/', restoreUser, asyncHandler(async (req, res) => {
   const { user } = req;
